Validates onBlur. Form validation useEffect updated to including greate range of conditions. Form validation context moved to new contact page, form moved into ui component.
