+++ src/app/lib/addContactFormElements.ts
@@ -1,0 +1,18 @@
+export const formElements = [
+  {
+    label: 'First Name',
+    name: 'firstName',
+  },
+  {
+    label: 'Last Name',
+    name: 'lastName',
+  },
+  {
+    label: 'E-mail Address',
+    name: 'email',
+  },
+  {
+    label: 'Phone Number',
+    name: 'phone',
+  },
+];
